requirements.txt
click
bs4
pandas
html5lib
lxml
pyyaml